+++ packages/core/ios/RNSentryEvents.m
@@ -1,3 +1,4 @@
 #import "RNSentryEvents.h"
 
 NSString *const RNSentryNewFrameEvent = @"rn_sentry_new_frame";
+NSString *const RNSentryOnShakeEvent = @"rn_sentry_on_shake";